+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### `@liveblocks/node`
 
 - Allow passing optional AbortSignal to all client methods
+- Fix bug in encoding of error information in the LiveblocksError when an API
+  call fails
 
 ## v2.18.3